+++ packages/flutter_riverpod/CHANGELOG.md
@@ -5,6 +5,20 @@
 
   For more info, see http://riverpod.dev/docs/concepts/modifiers/auto_dispose#the-argument-type-autodisposeprovider-cant-be-assigned-to-the-parameter-type-alwaysaliveproviderbase
 
+- `ScopedProvider` now accepts `null` as a function:
+
+  ```dart
+  final example = ScopedProvider<int>(null);
+  ```
+
+  Which is equivalent to:
+
+  ```dart
+  final example = ScopedProvider<int>((watch) => throw UnimplementedError(''));
+  ```
+
+- Fixed a bug where `context.refresh` may not work properly if the widget tree
+  contains multiple `ProviderScope`.
 
 # 0.6.1